requirements.txt
textual>=0.40.0
psutil>=5.9.0
toml>=0.10.2
click>=8.1.0
python-daemon>=3.0.0
yara-python>=4.3.0
python-magic>=0.4.27
pyelftools>=0.29
aiosqlite>=0.19.0
